+++ src/main/java/com/spotify/docker/client/DockerClient.java
@@ -71,7 +71,12 @@ import com.spotify.docker.client.messages.swarm.SecretSpec;
 import com.spotify.docker.client.messages.swarm.Service;
 import com.spotify.docker.client.messages.swarm.ServiceSpec;
 import com.spotify.docker.client.messages.swarm.Swarm;
+import com.spotify.docker.client.messages.swarm.SwarmInit;
+import com.spotify.docker.client.messages.swarm.SwarmJoin;
+import com.spotify.docker.client.messages.swarm.SwarmSpec;
 import com.spotify.docker.client.messages.swarm.Task;
+import com.spotify.docker.client.messages.swarm.UnlockKey;
+
 import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
@@ -283,47 +288,6 @@ public interface DockerClient extends Closeable {
   void load(String image, InputStream imagePayload, ProgressHandler handler)
       throws DockerException, InterruptedException;
 
-
-  /**
-   * Creates a single image from a tarball. This method also tags the image
-   * with the given image name upon loading completion.
-   *
-   * @param image        the name to assign to the image.
-   * @param imagePayload the image's payload (i.e.: the stream corresponding to the image's .tar
-   *                     file).
-   * @param registryAuth the {@link RegistryAuth} needed to pull the image.
-   * @throws DockerException      if a server error occurred (500).
-   * @throws InterruptedException if the thread is interrupted.
-   *
-   * @deprecated Use {@link #load(InputStream)} to load a set of image layers from a tarball. Use
-   * {@link #create(String, InputStream)} to create a single image from the contents of a tarball.
-   */
-  @Deprecated
-  void load(String image, InputStream imagePayload, RegistryAuth registryAuth)
-      throws DockerException, InterruptedException;
-
-
-  /**
-   * Creates a single image from a tarball. This method also tags the image
-   * with the given image name upon loading completion.
-   *
-   * @param image        the name to assign to the image.
-   * @param imagePayload the image's payload (i.e.: the stream corresponding to the image's .tar
-   *                     file).
-   * @param registryAuth the {@link RegistryAuth} needed to pull the image.
-   * @param handler      The handler to use for processing each progress message received from
-   *                     Docker.
-   * @throws DockerException      if a server error occurred (500).
-   * @throws InterruptedException if the thread is interrupted.
-   *
-   * @deprecated Use {@link #load(InputStream)} to load a set of image layers from a tarball. Use
-   *             {@link #create(String, InputStream, ProgressHandler)} to create a single image from
-   *             the contents of a tarball.
-   */
-  @Deprecated
-  void load(String image, InputStream imagePayload, RegistryAuth registryAuth,
-            ProgressHandler handler) throws DockerException, InterruptedException;
-
   /**
    * Load a set of images and tags from a tarball.
    *
@@ -391,26 +355,6 @@ public interface DockerClient extends Closeable {
    * @throws InterruptedException if the thread is interrupted.
    */
   InputStream save(String... images) throws DockerException, IOException, InterruptedException;
-
-  /**
-   * Get a tarball containing all images and metadata for the repository specified.
-   * @param image the name or id of the image to save. If a specific name and tag
-   *              (e.g. ubuntu:latest), then only that image (and its parents) are returned.
-   *              If an image ID, similarly only that image (and its parents) are returned,
-   *              but with the exclusion of the 'repositories' file in the tarball,
-   *              as there were no image names referenced.
-   * @param registryAuth the {@link RegistryAuth} needed to pull the image.
-   * @return the image's .tar stream.
-   * @throws DockerException      if a server error occurred (500).
-   * @throws IOException          if the server started returning, but an I/O error occurred in the
-   *                              context of processing it on the client-side.
-   * @throws InterruptedException if the thread is interrupted.
-   *
-   * @deprecated RegistryAuth is not required. Use {@link #save(String...)}.
-   */
-  @Deprecated
-  InputStream save(String image, RegistryAuth registryAuth)
-      throws DockerException, IOException, InterruptedException;
 
   /**
    * Get a tarball containing all images and metadata for one or more repositories.
@@ -1482,13 +1426,128 @@ public interface DockerClient extends Closeable {
       throws DockerException, InterruptedException;
 
   /**
-   * Inspect the Swarm cluster. Only available in Docker API &gt;= 1.24.
+   * Inspect the swarm. Only available in Docker API &gt;= 1.24.
    *
    * @return Info about a swarm
-   * @throws DockerException      if a server error occurred (500)
+   * @throws DockerException      If a server error occurred (500)
    * @throws InterruptedException If the thread is interrupted
    */
   Swarm inspectSwarm() throws DockerException, InterruptedException;
+
+  /**
+   * Initialize a new swarm. Only available in Docker API &gt;= 1.24.
+   *
+   * @return Node ID
+   * @throws DockerException      If a server error occurred (500)
+   * @throws InterruptedException If the thread is interrupted
+   */
+  String initSwarm(SwarmInit swarmInit) throws DockerException, InterruptedException;
+
+  /**
+   * Join an existing swarm. Only available in Docker API &gt;= 1.24.
+   *
+   * @throws DockerException      If a server error occurred (500)
+   * @throws InterruptedException If the thread is interrupted
+   */
+  void joinSwarm(SwarmJoin swarmJoin) throws DockerException, InterruptedException;
+
+  /**
+   * Leave a swarm. Only available in Docker API &gt;= 1.24.
+   *
+   * @throws DockerException      If a server error occurred (500)
+   * @throws InterruptedException If the thread is interrupted
+   */
+  void leaveSwarm() throws DockerException, InterruptedException;
+
+  /**
+   * Leave a swarm forcefully. Only available in Docker API &gt;= 1.24.
+   * Force leave swarm, even if this is the last manager or if leaving will break the cluster.
+   *
+   * @param force                 Whether to leave forcefully
+   * @throws DockerException      If a server error occurred (500)
+   * @throws InterruptedException If the thread is interrupted
+   */
+  void leaveSwarm(boolean force) throws DockerException, InterruptedException;
+
+  /**
+   * Update a swarm. Only available in Docker API &gt;= 1.24.
+   *
+   * @param version                The version number of the swarm object being updated.
+   *                               This is required to avoid conflicting writes.
+   * @param rotateWorkerToken      Set to true to rotate the worker join token.
+   * @param rotateManagerToken     Set to true to rotate the worker join token.
+   * @param rotateManagerUnlockKey Set to true to rotate the manager unlock key.
+   * @param spec                   {@link SwarmSpec}
+   * @throws DockerException       If a server error occurred (500)
+   * @throws InterruptedException  If the thread is interrupted
+   */
+  void updateSwarm(Long version,
+                   boolean rotateWorkerToken,
+                   boolean rotateManagerToken,
+                   boolean rotateManagerUnlockKey,
+                   SwarmSpec spec)
+      throws DockerException, InterruptedException;
+
+  /**
+   * Update a swarm. Only available in Docker API &gt;= 1.24.
+   *
+   * @param version                The version number of the swarm object being updated.
+   *                               This is required to avoid conflicting writes.
+   * @param rotateWorkerToken      Set to true to rotate the worker join token.
+   * @param rotateManagerToken     Set to true to rotate the worker join token.
+   * @param spec                   {@link SwarmSpec}
+   * @throws DockerException      If a server error occurred (500)
+   * @throws InterruptedException If the thread is interrupted
+   */
+  void updateSwarm(Long version,
+                   boolean rotateWorkerToken,
+                   boolean rotateManagerToken,
+                   SwarmSpec spec)
+      throws DockerException, InterruptedException;
+
+  /**
+   * Update a swarm. Only available in Docker API &gt;= 1.24.
+   *
+   * @param version                The version number of the swarm object being updated.
+   *                               This is required to avoid conflicting writes.
+   * @param rotateWorkerToken      Set to true to rotate the worker join token.
+   * @param spec                   {@link SwarmSpec}
+   * @throws DockerException      If a server error occurred (500)
+   * @throws InterruptedException If the thread is interrupted
+   */
+  void updateSwarm(Long version,
+                   boolean rotateWorkerToken,
+                   SwarmSpec spec)
+      throws DockerException, InterruptedException;
+
+  /**
+   * Update a swarm. Only available in Docker API &gt;= 1.24.
+   *
+   * @param version                The version number of the swarm object being updated.
+   *                               This is required to avoid conflicting writes.
+   * @param spec                   {@link SwarmSpec}
+   * @throws DockerException      If a server error occurred (500)
+   * @throws InterruptedException If the thread is interrupted
+   */
+  void updateSwarm(Long version, SwarmSpec spec) throws DockerException, InterruptedException;
+
+  /**
+   * Get an unlock key for unlocking a swarm. Only available in Docker API &gt;= 1.25.
+   *
+   * @return {@link UnlockKey}
+   * @throws DockerException      If a server error occurred (500)
+   * @throws InterruptedException If the thread is interrupted
+   */
+  UnlockKey unlockKey() throws DockerException, InterruptedException;
+
+  /**
+   * Unlock a swarm. Only available in Docker API &gt;= 1.25.
+   *
+   * @param unlockKey             {@link UnlockKey}
+   * @throws DockerException      If a server error occurred (500)
+   * @throws InterruptedException If the thread is interrupted
+   */
+  void unlock(UnlockKey unlockKey) throws DockerException, InterruptedException;
 
   /**
    * Create a new service. Only available in Docker API &gt;= 1.24.