+++ src/main/java/com/spotify/docker/client/messages/swarm/UnlockKey.java
@@ -28,34 +28,16 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.auto.value.AutoValue;
 
-import javax.annotation.Nullable;
-
 
 @AutoValue
 @JsonAutoDetect(fieldVisibility = ANY, getterVisibility = NONE, setterVisibility = NONE)
-public abstract class OrchestrationConfig {
+public abstract class UnlockKey {
 
-  @Nullable
-  @JsonProperty("TaskHistoryRetentionLimit")
-  public abstract Integer taskHistoryRetentionLimit();
+  @JsonProperty("UnlockKey")
+  public abstract String unlockKey();
 
   @JsonCreator
-  static OrchestrationConfig create(
-      @JsonProperty("TaskHistoryRetentionLimit") final Integer taskHistoryRetentionLimit) {
-    return builder()
-        .taskHistoryRetentionLimit(taskHistoryRetentionLimit)
-        .build();
-  }
-
-  @AutoValue.Builder
-  public abstract static class Builder {
-
-    public abstract Builder taskHistoryRetentionLimit(Integer taskHistoryRetentionLimit);
-
-    public abstract OrchestrationConfig build();
-  }
-
-  public static OrchestrationConfig.Builder builder() {
-    return new AutoValue_OrchestrationConfig.Builder();
+  static UnlockKey create(@JsonProperty("UnlockKey") final String unlockKey) {
+    return new AutoValue_UnlockKey(unlockKey);
   }
 }